crams-apps/requirements.txt
python-dateutil==2.8.1
pytz==2020.4
asgiref==3.2.10
Django==3.1.1
djangorestframework==3.11.1
pyyaml==5.4.1
uritemplate==3.0.1
